Official Documentation for the Hiect APIs and Streams.
Official Announcements regarding changes, downtime, etc. to the API and Streams will be reported.
Streams, endpoints, parameters, payloads, etc. decscribed in the documents in this repository are considered official and supported.
The use of any other streams, endpoints, parameters, or payloads, etc. is not supported; use them at your own risk and with no guarantees.